+++ packages/flutter_fixtures_core/lib/flutter_fixtures_core.dart
@@ -11,5 +11,6 @@ export 'src/fixture_selector.dart';
 export 'src/fixture_document.dart';
 export 'src/fixture_collection.dart';
 export 'src/fixture_selection_memory.dart';
+export 'src/http_file_fixture_source.dart';
 export 'src/http_fixture_source.dart';
 export 'src/openapi_fixture_source.dart';